requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
python-dotenv==1.0.0
psycopg2==2.9.9
gunicorn==21.2.0 
pytest_factoryboy==2.6.0